requirements.txt
flask
yolov5
rdflib
sparqlwrapper 
owlready2 
pyshacl
Pillow